+++ cohere/responses/classify.py
@@ -1,5 +1,6 @@
 from typing import Any, Dict, List, NamedTuple, Optional
 
+from cohere.logging import logger
 from cohere.responses.base import CohereObject
 
 LabelPrediction = NamedTuple("LabelPrediction", [("confidence", float)])
@@ -8,19 +9,50 @@ Example = NamedTuple("Example", [("text", str), ("label", str)])
 
 class Classification(CohereObject):
     def __init__(
-        self, input: str, prediction: str, confidence: float, labels: Dict[str, LabelPrediction], *args, **kwargs
+        self,
+        input: str,
+        predictions: Optional[List[str]],
+        confidences: Optional[List[float]],
+        prediction: Optional[str],
+        confidence: Optional[float],
+        labels: Dict[str, LabelPrediction],
+        classification_type: str,
+        *args,
+        **kwargs,
     ) -> None:
         super().__init__(*args, **kwargs)
         self.input = input
-        self.prediction = prediction
-        self.confidence = confidence
+        self._prediction = prediction  # to be removed
+        self._confidence = confidence  # to be removed
+        self.predictions = predictions
+        self.confidences = confidences
         self.labels = labels
+        self.classification_type = classification_type
+
+        if self._prediction is None or self._confidence is None:
+            if self._prediction is not None or self._confidence is not None:
+                raise ValueError("Cannot have one of `prediction` and `confidence` be None and not the other one")
+            if self.predictions is None or self.confidences is None:
+                raise ValueError("Cannot have `predictions` or `confidences` be None if `prediction` is None")
 
     def __repr__(self) -> str:
-        prediction = self.prediction
-        confidence = self.confidence
-        labels = self.labels
-        return f'Classification<prediction: "{prediction}", confidence: {confidence}, labels: {labels}>'
+        if self._prediction is not None:
+            return f'Classification<prediction: "{self._prediction}", confidence: {self._confidence}, labels: {self.labels}>'
+        else:
+            return f'Classification<predictions: "{self.predictions}", confidences: {self.confidences}, labels: {self.labels}>'
+
+    @property
+    def prediction(self):
+        logger.warning("`prediction` is deprecated and will be removed soon. Please use `predictions` instead.")
+        return self._prediction
+
+    @property
+    def confidence(self):
+        logger.warning("`confidence` is deprecated and will be removed soon. Please use `confidences` instead.")
+        return self._confidence
+
+    def is_multilabel(self) -> bool:
+        return self.classification_type == "multi-label"
 
 
 class Classifications(CohereObject):