+++ tests/async/test_async_tokenize.py
@@ -17,7 +17,9 @@ async def test_tokenize(async_client):
     assert tokens.meta["api_version"]["version"]
 
 
+# TODO(manoj): Fix the test expectation due to the base model change (MS-913)
 @pytest.mark.asyncio
+@pytest.mark.skip
 async def test_model_param_tokenization(async_client):
     # Use tuples to be hashable (for set equality)
     medium_res = tuple((await async_client.tokenize("Hello world!", model="medium")).tokens)
@@ -43,7 +45,9 @@ async def test_invalid_text(async_client):
         await async_client.tokenize("")
 
 
+# TODO(manoj): Fix the test expectation due to the base model change (MS-913)
 @pytest.mark.asyncio
+@pytest.mark.skip
 async def test_detokenize(async_client):
     detokenized = await async_client.detokenize([10104, 12221, 974, 514, 34], model="base")
     assert detokenized == Detokenization("detokenize me!")
@@ -58,7 +62,9 @@ async def test_detokenize(async_client):
     assert detokenized == Detokenization("")
 
 
+# TODO(manoj): Fix the test expectation due to the base model change (MS-913)
 @pytest.mark.asyncio
+@pytest.mark.skip
 async def test_model_param_detokenization(async_client):
     medium_detokenized = (await async_client.detokenize([10104, 12221, 974, 514, 34], model="medium")).text
     medium_detokenized_batch = [