+++ tests/sync/test_detokenize.py
@@ -1,5 +1,6 @@
 import unittest
 
+import pytest
 from utils import get_api_key
 
 import cohere
@@ -8,6 +9,8 @@ co = cohere.Client(get_api_key())
 
 
 class TestDetokenize(unittest.TestCase):
+    # TODO(manoj): Fix the test expectation due to the base model change (MS-913)
+    @pytest.mark.skip
     def test_detokenize_success(self):
         resp = co.detokenize([10104, 12221, 974, 514, 34], model="base")
         text = resp.text
@@ -16,6 +19,8 @@ class TestDetokenize(unittest.TestCase):
         self.assertTrue(resp.meta["api_version"])
         self.assertTrue(resp.meta["api_version"]["version"])
 
+    # TODO(manoj): Fix the test expectation due to the base model change (MS-913)
+    @pytest.mark.skip
     def test_detokenize_batched(self):
         _batch_size = 3
         texts = co.batch_detokenize([[10104, 12221, 974, 514, 34]] * _batch_size, model="base")
@@ -24,6 +29,8 @@ class TestDetokenize(unittest.TestCase):
             results.append(str(text))
         self.assertEqual(results, ["detokenize me!"] * _batch_size)
 
+    # TODO(manoj): Fix the test expectation due to the base model change (MS-913)
+    @pytest.mark.skip
     def test_empty_tokens(self):
         text = co.detokenize([]).text
         self.assertEqual(text, "")